fixed bug that epsilon is passed as int

+++ headers/ppo.h
@@ -31,7 +31,7 @@ void free_ppo(PPO* ppo);
 void collect_trajectories(TrajectoryBuffer* buffer, Env* env, GaussianPolicy* policy, int steps);
 void compute_gae(NeuralNetwork* V, TrajectoryBuffer* buffer, float gamma, float lambda);
 
-float policy_loss_and_grad(float* grad_logprob, float* grad_entropy, float* adv, float* logprobs,  float* old_logprobs, float entropy, float ent_coeff, int epsilon, int m);
+float policy_loss_and_grad(float* grad_logprob, float* grad_entropy, float* adv, float* logprobs,  float* old_logprobs, float entropy, float ent_coeff, float epsilon, int m);
 
 void train_ppo_epoch(PPO* ppo, int steps_per_epoch, int batch_size, int n_epochs_policy, int n_epochs_value);
 void eval_ppo(PPO* ppo, int steps);
